% Auto-generated: do not edit by hand
\name{dashWebsocket}

\alias{dashWebsocket}

\title{DashWebsocket component}

\description{
DashWebsocket is an adapter for websocket. It takes two property, `port` and 'msg' `port` indicates the websocket url port `msg` display the message returns from webscoket
}

\usage{
dashWebsocket(id=NULL, msg=NULL, port=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{msg}{Character. The websocket response message.}

\item{port}{Character. The port for websocket.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

