% Auto-generated: do not edit by hand
\name{websocket}

\alias{websocket}

\title{Websocket component}

\description{

}

\usage{
websocket(id=NULL, message=NULL, send=NULL, url=NULL)
}

\arguments{
\item{id}{Character. }

\item{message}{Named list | character. }

\item{send}{Named list | character. }

\item{url}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

