% Auto-generated: do not edit by hand
\name{dashWebsocket}

\alias{dashWebsocket}

\title{DashWebsocket component}

\description{
DashWebsocket is an adapter for websocket. It takes two property, `url` and 'msg' `url` indicates the websocket url `msg` display the message returns from webscoket
}

\usage{
dashWebsocket(id=NULL, msg=NULL, url=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{msg}{Character. The websocket response message.}

\item{url}{Character. The url for websocket.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

